Migrate additional entry manager tests to unit project
